Pocket uart expand (#434)
* Modified pocket uart expand logic

* Modified pocket uart expand logic
