+++ apps/tuya_t5_pocket/tuya_t5_pocket_ai/CMakeLists.txt
@@ -48,6 +48,7 @@ target_compile_options(${EXAMPLE_LIB}
 ########################################
 
 add_subdirectory(${APP_PATH}/src/display)
+add_subdirectory(${APP_PATH}/src/expand)
 
 add_subdirectory(${APP_PATH}/../../tuya.ai/ai_components/ai_audio)
 target_include_directories(${EXAMPLE_LIB} PRIVATE ${APP_PATH}/../../tuya.ai/ai_components/ai_audio)
