+++ apps/tuya_t5_pocket/tuya_t5_pocket_ai/src/expand/inc/u2g_tbl.inc
